refactor(liferay-cli): move utils to top level directory
So that we can leverage them from adapt process.
