+++ projects/js-toolkit/packages/liferay-cli/src/new/facet-configuration/index.ts
@@ -11,8 +11,8 @@ import {
 	transformTextFile,
 } from '@liferay/js-toolkit-core';
 
-import ensureOutputFile from '../util/ensureOutputFile';
-import prompt from '../util/prompt';
+import ensureOutputFile from '../../util/ensureOutputFile';
+import prompt from '../../util/prompt';
 
 import type {Facet, Options} from '../index';
