+++ projects/js-toolkit/packages/liferay-cli/src/new/facet-react/index.ts
@@ -12,7 +12,7 @@ import {
 	transformTextFile,
 } from '@liferay/js-toolkit-core';
 
-import ensureOutputFile from '../util/ensureOutputFile';
+import ensureOutputFile from '../../util/ensureOutputFile';
 
 import type {Facet, Options} from '../index';
