+++ projects/js-toolkit/packages/liferay-cli/src/new/facet-sample-configuration/index.ts
@@ -10,7 +10,7 @@ import {
 	transformTextFile,
 } from '@liferay/js-toolkit-core';
 
-import ensureOutputFile from '../util/ensureOutputFile';
+import ensureOutputFile from '../../util/ensureOutputFile';
 
 import type {Facet, Options} from '../index';
