+++ projects/js-toolkit/packages/liferay-cli/src/new/facet-sample-styles/index.ts
@@ -9,7 +9,7 @@ import {
 	transformTextFile,
 } from '@liferay/js-toolkit-core';
 
-import ensureOutputFile from '../util/ensureOutputFile';
+import ensureOutputFile from '../../util/ensureOutputFile';
 
 import type {Facet, Options} from '../index';
