+++ projects/js-toolkit/packages/liferay-cli/src/util/getPortletName.ts
