+++ projects/js-toolkit/packages/liferay-cli/src/util/toHumanReadable.ts
